requirements.txt
asgiref==3.2.5
autopep8==1.5.4
Babel==2.8.0
beautifulsoup4==4.9.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-appconf==1.0.3
django-avatar==5.0.0
django-crispy-forms==1.9.2
django-drf-filepond==0.3.0
django-filepond==0.1.2
django-jsonfield==1.4.1
django-mathfilters==1.0.0
django-model-utils==4.0.0
django-multiselectfield==0.1.12
django-multiupload==0.5.2
django-notifications-hq==1.6.0
django-notify-x==0.1.9
django-phonenumber-field==4.0.0
django-shortuuidfield==0.1.3
django-storages==1.10.1
djangorestframework==3.12.2
docopt==0.6.2
Faker==4.14.0
idna==2.10
Js2Py==0.70
jsonfield==3.1.0
olefile==0.46
packaging==20.4
phonenumbers==8.11.5
PIL @ file:///C:/Users/Samad%20Farooq/pipwin/PIL-2.0%2Bdummy-py2.py3-none-any.whl
pillow @ file:///C:/Users/Samad%20Farooq/pipwin/Pillow-8.0.1-cp39-cp39-win_amd64.whl
pinax-notifications==6.0.0
pipwin==0.5.0
psycopg2 @ file:///C:/Users/Samad%20Farooq/pipwin/psycopg2-2.8.6-cp39-cp39-win_amd64.whl
pycodestyle==2.6.0
pyjsparser==2.7.1
pyparsing==2.4.7
PyPrind==2.11.2
pySmartDL==1.3.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.24.0
shortuuid==1.0.1
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
swapper==1.1.2.post1
text-unidecode==1.3
toml==0.10.2
tzlocal==2.1
urllib3==1.25.11
